feat : 모든 도메인 JpaRepository 세팅
