+++ src/main/java/com/hyperlink/server/domain/attentionCategory/domain/AttentionCategoryRepository.java
@@ -1,0 +1,8 @@
+package com.hyperlink.server.domain.attentionCategory.domain;
+
+import com.hyperlink.server.domain.attentionCategory.domain.entity.AttentionCategory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AttentionCategoryRepository extends JpaRepository<AttentionCategory, Long> {
+
+}
