+++ src/main/java/com/hyperlink/server/domain/category/domain/CategoryRepository.java
@@ -1,0 +1,8 @@
+package com.hyperlink.server.domain.category.domain;
+
+import com.hyperlink.server.domain.category.domain.entity.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+
+}
