+++ src/main/java/com/hyperlink/server/domain/contentCategory/domain/ContentCategoryRepository.java
@@ -1,0 +1,8 @@
+package com.hyperlink.server.domain.contentCategory.domain;
+
+import com.hyperlink.server.domain.contentCategory.domain.entity.ContentCategory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ContentCategoryRepository extends JpaRepository<ContentCategory, Long> {
+
+}
