+++ src/main/java/com/hyperlink/server/domain/contentCategory/domain/entity/ContentCategory.java
@@ -1,8 +1,8 @@
-package com.hyperlink.server.domain.contentCategory.entity;
+package com.hyperlink.server.domain.contentCategory.domain.entity;
 
 import com.hyperlink.server.domain.category.domain.entity.Category;
 import com.hyperlink.server.domain.common.BaseEntity;
-import com.hyperlink.server.domain.content.entity.Content;
+import com.hyperlink.server.domain.content.domain.entity.Content;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
