+++ src/main/java/com/hyperlink/server/domain/creator/domain/CreatorRepository.java
@@ -1,0 +1,8 @@
+package com.hyperlink.server.domain.creator.domain;
+
+import com.hyperlink.server.domain.creator.domain.entity.Creator;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CreatorRepository extends JpaRepository<Creator, Long> {
+
+}
