+++ src/main/java/com/hyperlink/server/domain/creatorCategory/domain/CreatorCategoryRepository.java
@@ -1,0 +1,8 @@
+package com.hyperlink.server.domain.creatorCategory.domain;
+
+import com.hyperlink.server.domain.creatorCategory.domain.entity.CreatorCategory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CreatorCategoryRepository extends JpaRepository<CreatorCategory, Long> {
+
+}
