+++ src/main/java/com/hyperlink/server/domain/creatorCategory/domain/entity/CreatorCategory.java
@@ -1,8 +1,7 @@
-package com.hyperlink.server.domain.subscription.entity;
+package com.hyperlink.server.domain.creatorCategory.domain.entity;
 
-import com.hyperlink.server.domain.common.BaseEntity;
-import com.hyperlink.server.domain.creator.entity.Creator;
-import com.hyperlink.server.domain.member.entity.Member;
+import com.hyperlink.server.domain.category.domain.entity.Category;
+import com.hyperlink.server.domain.creator.domain.entity.Creator;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -18,18 +17,19 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Subscription extends BaseEntity {
+public class CreatorCategory {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "subscription_id")
+  @Column(name = "creator_category_id")
   private Long id;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "member_id")
-  private Member member;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "creator_id")
   private Creator creator;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "category_id")
+  private Category category;
+
 }
